RF_00.004.13 | | New item > Create New Folder > Verify entering special characters as the Folder name appears error message -> POM, moved to FolderTest
